app/requirements.txt
flask-restx==1.3.0
Flask-JWT-Extended==4.6.0
flask==2.2.5
flask_sqlalchemy==3.1.1
openpyxl==3.1.2
python-dotenv==1.0.1
flask-Cors==4.0.0
pandas==2.2.1
gunicorn